Replace query params with batching circuits + use batching public inputs for tabular queries
