+++ verifiable-db/src/query/batching/circuits/row_chunk_processing.rs
@@ -10,34 +10,27 @@ use recursion_framework::circuit_builder::CircuitLogicWires;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 
 use crate::query::{
-    aggregation::QueryBounds,
-    batching::{
-        public_inputs::PublicInputs,
-        row_chunk,
+    aggregation::QueryBounds, batching::row_chunk::
+    {
         row_process_gadget::{RowProcessingGadgetInputWires, RowProcessingGadgetInputs},
-    },
-    computational_hash_ids::ColumnIDs,
-    universal_circuit::{
+        aggregate_chunks::aggregate_chunks, RowChunkDataTarget,
+    }, 
+        computational_hash_ids::ColumnIDs, pi_len, public_inputs::PublicInputs, universal_circuit::{
         universal_circuit_inputs::{BasicOperation, Placeholders, ResultStructure},
         universal_query_gadget::{
             OutputComponent, UniversalQueryHashInputWires, UniversalQueryHashInputs,
         },
-    },
+    }
 };
 
 use mp2_common::{
     public_inputs::PublicInputCommon,
     serialization::{deserialize_long_array, serialize_long_array},
-    u256::{CircuitBuilderU256, UInt256Target, WitnessWriteU256},
     utils::ToTargets,
     D, F,
 };
 
-use self::row_chunk::{aggregate_chunks::aggregate_chunks, RowChunkDataTarget};
-
 use anyhow::{ensure, Result};
-
-use super::api::num_io;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct RowChunkProcessingWires<
@@ -67,8 +60,6 @@ pub struct RowChunkProcessingWires<
         MAX_NUM_RESULTS,
         T,
     >,
-    min_query_primary: UInt256Target,
-    max_query_primary: UInt256Target,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -193,15 +184,11 @@ where
         T,
     > {
         let query_input_wires = UniversalQueryHashInputs::build(b);
-        let [min_query_primary, max_query_primary] = b.add_virtual_u256_arr_unsafe(); // unsafe should be ok since
-                                                                                      // we are exposing these values as public inputs
         let first_row_wires = RowProcessingGadgetInputs::build(
             b,
             &query_input_wires.input_wires,
             &query_input_wires.min_secondary,
             &query_input_wires.max_secondary,
-            &min_query_primary,
-            &max_query_primary,
         );
         // enforce first row is non-dummy
         b.assert_one(
@@ -222,8 +209,6 @@ where
                 &query_input_wires.input_wires,
                 &query_input_wires.min_secondary,
                 &query_input_wires.max_secondary,
-                &min_query_primary,
-                &max_query_primary,
             );
             row_inputs.push(RowProcessingGadgetInputWires::from(&row_wires));
             let is_second_non_dummy = row_wires.value_wires.input_wires.is_non_dummy_row;
@@ -232,7 +217,7 @@ where
                 b,
                 &chunk,
                 &current_chunk,
-                (&min_query_primary, &max_query_primary),
+                (&query_input_wires.input_wires.min_query_primary, &query_input_wires.input_wires.max_query_primary),
                 (
                     &query_input_wires.min_secondary,
                     &query_input_wires.max_secondary,
@@ -258,8 +243,8 @@ where
             &query_input_wires.agg_ops_ids,
             &row_chunk.left_boundary_row.to_targets(),
             &row_chunk.right_boundary_row.to_targets(),
-            &min_query_primary.to_targets(),
-            &max_query_primary.to_targets(),
+            &query_input_wires.input_wires.min_query_primary.to_targets(),
+            &query_input_wires.input_wires.max_query_primary.to_targets(),
             &query_input_wires.min_secondary.to_targets(),
             &query_input_wires.max_secondary.to_targets(),
             &[overflow.target],
@@ -271,8 +256,6 @@ where
         RowChunkProcessingWires {
             row_inputs: row_inputs.try_into().unwrap(),
             universal_query_inputs: query_input_wires.input_wires,
-            min_query_primary,
-            max_query_primary,
         }
     }
 
@@ -296,12 +279,6 @@ where
             .for_each(|(value, target)| value.assign(pw, target));
         self.universal_query_inputs
             .assign(pw, &wires.universal_query_inputs);
-        [
-            (self.min_query_primary, &wires.min_query_primary),
-            (self.max_query_primary, &wires.max_query_primary),
-        ]
-        .into_iter()
-        .for_each(|(value, target)| pw.set_u256_target(target, value));
     }
 
     /// This method returns the ids of the placeholders employed to compute the placeholder hash,
@@ -354,7 +331,7 @@ where
         T,
     >;
 
-    const NUM_PUBLIC_INPUTS: usize = num_io::<MAX_NUM_RESULTS>();
+    const NUM_PUBLIC_INPUTS: usize = pi_len::<MAX_NUM_RESULTS>();
 
     fn circuit_logic(
         builder: &mut CircuitBuilder<F, D>,
@@ -405,10 +382,12 @@ mod tests {
                 row_chunk_processing::RowChunkProcessingCircuit,
                 tests::{build_test_tree, compute_output_values_for_row},
             },
-            public_inputs::PublicInputs,
-            row_chunk::tests::{BoundaryRowData, BoundaryRowNodeInfo},
-            row_process_gadget::RowProcessingGadgetInputs,
+            row_chunk::{
+                tests::{BoundaryRowData, BoundaryRowNodeInfo},
+                row_process_gadget::RowProcessingGadgetInputs
+            },
         },
+        public_inputs::PublicInputs,
         computational_hash_ids::{
             AggregationOperation, ColumnIDs, Identifiers, Operation, PlaceholderIdentifier,
         },
