+++ verifiable-db/src/query/batching/mod.rs
@@ -1,6 +1,2 @@
 pub(crate) mod circuits;
-pub mod public_inputs;
 pub(crate) mod row_chunk;
-mod row_process_gadget;
-
-pub use circuits::api::*;
