+++ verifiable-db/src/query/mod.rs
@@ -1,4 +1,4 @@
-use mp2_common::F;
+use plonky2::iop::target::Target;
 use public_inputs::PublicInputs;
 
 pub mod aggregation;
@@ -10,5 +10,5 @@ pub mod public_inputs;
 pub mod universal_circuit;
 
 pub const fn pi_len<const MAX_NUM_RESULTS: usize>() -> usize {
-    PublicInputs::<F, MAX_NUM_RESULTS>::total_len()
+    PublicInputs::<Target, MAX_NUM_RESULTS>::total_len()
 }
