+++ verifiable-db/src/revelation/revelation_without_results_tree.rs
@@ -3,12 +3,7 @@
 use crate::{
     ivc::PublicInputs as OriginalTreePublicInputs,
     query::{
-        computational_hash_ids::AggregationOperation,
-        public_inputs::PublicInputs as QueryProofPublicInputs,
-        universal_circuit::{
-            universal_query_gadget::OutputValuesTarget, ComputationalHashTarget,
-            MembershipHashTarget, PlaceholderHashTarget,
-        },
+        public_inputs::PublicInputs as QueryProofPublicInputs, computational_hash_ids::AggregationOperation, pi_len as query_pi_len,
     },
     revelation::PublicInputs,
 };
@@ -23,7 +18,7 @@ use mp2_common::{
     serialization::{deserialize, serialize},
     types::CBuilder,
     u256::{CircuitBuilderU256, UInt256Target},
-    utils::{FromTargets, ToTargets},
+    utils::ToTargets,
     C, D, F,
 };
 use plonky2::{
@@ -47,15 +42,10 @@ use recursion_framework::{
 use serde::{Deserialize, Serialize};
 
 use super::{
-    num_query_io, num_query_io_no_results_tree, pi_len as revelation_pi_len,
+    pi_len as revelation_pi_len,
     placeholders_check::{CheckPlaceholderGadget, CheckPlaceholderInputWires},
     NUM_PREPROCESSING_IO,
 };
-
-#[cfg(feature = "batching_circuits")]
-use super::batching::RevelationCircuitBatching;
-#[cfg(feature = "batching_circuits")]
-use crate::query::batching::public_inputs::PublicInputs as BatchingPublicInputs;
 
 // L: maximum number of results
 // S: maximum number of items in each result
@@ -81,46 +71,6 @@ pub struct RevelationWithoutResultsTreeCircuit<
     pub(crate) check_placeholder: CheckPlaceholderGadget<PH, PP>,
 }
 
-/// Data structure containing the wires corresponding to public inputs
-/// of the query proof employed in the revelation circuit. It is a
-/// data structure employed to represent the public inputs of the query
-/// proof in a revelation circuit both for query proofs generated with
-/// batching circuits, and for query proofs generated with non-batching
-/// circuits
-pub(crate) struct QueryProofInputWires<const S: usize>
-where
-    [(); S - 1]:,
-{
-    pub(crate) tree_hash: MembershipHashTarget,
-    pub(crate) results: OutputValuesTarget<S>,
-    pub(crate) entry_count: Target,
-    pub(crate) overflow: Target,
-    pub(crate) placeholder_hash: PlaceholderHashTarget,
-    pub(crate) computational_hash: ComputationalHashTarget,
-    pub(crate) min_primary: UInt256Target,
-    pub(crate) max_primary: UInt256Target,
-    pub(crate) ops: [Target; S],
-}
-
-impl<'a, const S: usize> From<&'a QueryProofPublicInputs<'a, Target, S>> for QueryProofInputWires<S>
-where
-    [(); S - 1]:,
-{
-    fn from(value: &'a QueryProofPublicInputs<Target, S>) -> Self {
-        Self {
-            tree_hash: value.tree_hash_target(),
-            results: OutputValuesTarget::from_targets(value.to_values_raw()),
-            entry_count: value.num_matching_rows_target(),
-            overflow: value.overflow_flag_target().target,
-            placeholder_hash: value.placeholder_hash_target(),
-            computational_hash: value.computational_hash_target(),
-            min_primary: value.min_query_target(),
-            max_primary: value.max_query_target(),
-            ops: value.operation_ids_target(),
-        }
-    }
-}
-
 impl<const L: usize, const S: usize, const PH: usize, const PP: usize>
     RevelationWithoutResultsTreeCircuit<L, S, PH, PP>
 where
@@ -133,18 +83,6 @@ where
         // proof of construction of the original tree in the pre-processing stage (IVC proof)
         original_tree_proof: &OriginalTreePublicInputs<Target>,
     ) -> RevelationWithoutResultsTreeWires<L, S, PH, PP> {
-        Self::build_core(b, query_proof.into(), original_tree_proof)
-    }
-
-    // Internal build method taking as input `QueryProofInputWires` instead of the public inputs of the
-    // query proof circuits without batching. The core logic is placed in this method because it is shared
-    // with the `RevelationCircuitBatchingCircuits` circuit
-    pub(crate) fn build_core(
-        b: &mut CBuilder,
-        query_proof: QueryProofInputWires<S>,
-        // proof of construction of the original tree in the pre-processing stage (IVC proof)
-        original_tree_proof: &OriginalTreePublicInputs<Target>,
-    ) -> RevelationWithoutResultsTreeWires<L, S, PH, PP> {
         let zero = b.zero();
         let u256_zero = b.zero_u256();
 
@@ -153,16 +91,16 @@ where
             .map(|op| b.constant(op.to_field()));
 
         // Convert the entry count to an Uint256.
-        let entry_count = UInt256Target::new_from_target(b, query_proof.entry_count);
+        let entry_count = UInt256Target::new_from_target(b, query_proof.num_matching_rows_target());
 
         // Compute the output results array, and deal with AVG and COUNT operations if any.
         let mut results = Vec::with_capacity(L * S);
         // flag to determine whether entry count is zero
         let is_entry_count_zero = b.add_virtual_bool_target_unsafe();
-        query_proof.ops.into_iter().enumerate().for_each(|(i, op)| {
+        query_proof.operation_ids_target().into_iter().enumerate().for_each(|(i, op)| {
             let is_op_avg = b.is_equal(op, op_avg);
             let is_op_count = b.is_equal(op, op_count);
-            let result = query_proof.results.value_target_at_index(i);
+            let result = query_proof.value_target_at_index(i);
 
             // Compute the AVG result (and it's set to zero if the divisor is zero).
             let (avg_result, _, is_divisor_zero) = b.div_u256(&result, &entry_count);
@@ -180,11 +118,11 @@ where
         // `check_placeholders` function:
         // H(pQ.H_p || pQ.MIN_I || pQ.MAX_I)
         let inputs = query_proof
-            .placeholder_hash
+            .placeholder_hash_target()
             .to_targets()
             .into_iter()
-            .chain(query_proof.min_primary.to_targets())
-            .chain(query_proof.max_primary.to_targets())
+            .chain(query_proof.min_primary_target().to_targets())
+            .chain(query_proof.max_primary_target().to_targets())
             .collect();
         let final_placeholder_hash = b.hash_n_to_hash_no_pad::<H>(inputs);
 
@@ -194,13 +132,13 @@ where
 
         // Check that the tree employed to build the queries is the same as the
         // tree constructed in pre-processing.
-        b.connect_hashes(query_proof.tree_hash, original_tree_proof.merkle_hash());
+        b.connect_hashes(query_proof.tree_hash_target(), original_tree_proof.merkle_hash());
 
         // Add the hash of placeholder identifiers and pre-processing metadata
         // hash to the computational hash:
         // H(pQ.C || placeholder_ids_hash || pQ.M)
         let inputs = query_proof
-            .computational_hash
+            .computational_hash_target()
             .to_targets()
             .iter()
             .chain(&check_placeholder_wires.placeholder_id_hash.to_targets())
@@ -221,6 +159,97 @@ where
 
         let flat_computational_hash = flatten_poseidon_hash_target(b, computational_hash);
 
+        // additional constraints on boundary rows to ensure completeness of proven rows
+        // (i.e., that we look at all the rows with primary and secondary index values in the query range)
+
+        let left_boundary_row = query_proof.left_boundary_row_target();
+
+        // 1. Either the index tree node of left boundary row has no predecessor, or
+        //    the value of the predecessor is smaller than MIN_primary
+        let smaller_than_min_primary = b.is_less_than_u256(
+            &left_boundary_row.index_node_info.predecessor_info.value,
+            &query_proof.min_primary_target(),
+        );
+        // assert not pQ.left_boundary_row.index_node_data.predecessor_info.is_found or
+        // pQ.left_boundary_row.index_node_data.predecessor_value < pQ.MIN_primary
+        let constraint = b.and(
+            left_boundary_row.index_node_info.predecessor_info.is_found,
+            smaller_than_min_primary,
+        );
+        b.connect(
+            left_boundary_row
+                .index_node_info
+                .predecessor_info
+                .is_found
+                .target,
+            constraint.target,
+        );
+
+        // 2. Either the rows tree node storing left boundary row has no predecessor, or
+        //    the value of the predecessor is smaller than MIN_secondary
+        let smaller_than_min_secondary = b.is_less_than_u256(
+            &left_boundary_row.row_node_info.predecessor_info.value,
+            &query_proof.min_secondary_target(),
+        );
+        // assert not pQ.left_boundary_row.row_node_data.predecessor_info.is_found or
+        // pQ.left_boundary_row.row_node_data.predecessor_value < pQ.MIN_secondary
+        let constraint = b.and(
+            left_boundary_row.row_node_info.predecessor_info.is_found,
+            smaller_than_min_secondary,
+        );
+        b.connect(
+            left_boundary_row
+                .row_node_info
+                .predecessor_info
+                .is_found
+                .target,
+            constraint.target,
+        );
+
+        let right_boundary_row = query_proof.right_boundary_row_target();
+
+        // 3. Either the index tree node of right boundary row has no successor, or
+        //    the value of the successor is greater than MAX_primary
+        let greater_than_max_primary = b.is_greater_than_u256(
+            &right_boundary_row.index_node_info.successor_info.value,
+            &query_proof.max_primary_target(),
+        );
+        // assert not pQ.right_boundary_row.index_node_data.successor_info.is_found or
+        // pQ.right_boundary_row.index_node_data.successor_value > pQ.MAX_primary
+        let constraint = b.and(
+            right_boundary_row.index_node_info.successor_info.is_found,
+            greater_than_max_primary,
+        );
+        b.connect(
+            right_boundary_row
+                .index_node_info
+                .successor_info
+                .is_found
+                .target,
+            constraint.target,
+        );
+
+        // 4. Either the rows tree node storing right boundary row has no successor, or
+        //    the value of the successor is greater than MAX_secondary
+        let greater_than_max_secondary = b.is_greater_than_u256(
+            &right_boundary_row.row_node_info.successor_info.value,
+            &query_proof.max_secondary_target(),
+        );
+        // assert not pQ.right_boundary_row.row_node_data.successor_info.is_found or
+        // pQ.right_boundary_row.row_node_data.successor_value > pQ.MAX_secondary
+        let constraint = b.and(
+            right_boundary_row.row_node_info.successor_info.is_found,
+            greater_than_max_secondary,
+        );
+        b.connect(
+            right_boundary_row
+                .row_node_info
+                .successor_info
+                .is_found
+                .target,
+            constraint.target,
+        );
+
         // Register the public innputs.
         PublicInputs::<_, L, S, PH>::new(
             &original_tree_proof.block_hash(),
@@ -230,8 +259,8 @@ where
             &[check_placeholder_wires.num_placeholders],
             // The aggregation query proof only has one result.
             &[num_results.target],
-            &[query_proof.entry_count],
-            &[query_proof.overflow],
+            &[query_proof.num_matching_rows_target()],
+            &[query_proof.overflow_flag_target().target],
             // Query limit
             &[zero],
             // Query offset
@@ -280,8 +309,7 @@ impl<const L: usize, const S: usize, const PH: usize, const PP: usize> CircuitLo
     for RecursiveCircuitWires<L, S, PH, PP>
 where
     [(); S - 1]:,
-    [(); num_query_io::<S>()]:,
-    [(); num_query_io_no_results_tree::<S>()]:,
+    [(); query_pi_len::<S>()]:,
     [(); <H as Hasher<F>>::HASH_SIZE]:,
 {
     type CircuitBuilderParams = CircuitBuilderParams;
@@ -299,7 +327,7 @@ where
             F,
             C,
             D,
-            { num_query_io_no_results_tree::<S>() },
+            { query_pi_len::<S>() },
         >::new(default_config(), &builder_parameters.query_circuit_set);
         let query_verifier = query_verifier.verify_proof_in_circuit_set(builder);
         let preprocessing_verifier =
@@ -313,19 +341,10 @@ where
         );
         let preprocessing_pi =
             OriginalTreePublicInputs::from_slice(&preprocessing_proof.public_inputs);
-        #[cfg(feature = "batching_circuits")]
-        let revelation_circuit = {
-            let query_pi = BatchingPublicInputs::from_slice(
-                query_verifier
-                    .get_public_input_targets::<F, { num_query_io_no_results_tree::<S>() }>(),
-            );
-            RevelationCircuitBatching::build(builder, &query_pi, &preprocessing_pi)
-        };
-        #[cfg(not(feature = "batching_circuits"))]
         let revelation_circuit = {
             let query_pi = QueryProofPublicInputs::from_slice(
                 query_verifier
-                    .get_public_input_targets::<F, { num_query_io_no_results_tree::<S>() }>(),
+                    .get_public_input_targets::<F, { query_pi_len::<S>() }>(),
             );
             RevelationWithoutResultsTreeCircuit::build(builder, &query_pi, &preprocessing_pi)
         };
@@ -349,27 +368,50 @@ where
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::{
-        query::{
-            public_inputs::QueryPublicInputs,
-            universal_circuit::universal_query_gadget::OutputValues,
-        },
-        revelation::tests::{compute_results_from_query_proof_outputs, TestPlaceholders},
-        test_utils::{
-            random_aggregation_operations, random_aggregation_public_inputs,
-            random_original_tree_proof,
-        },
-    };
+    use std::array;
+
     use alloy::primitives::U256;
+    use itertools::Itertools;
     use mp2_common::{
-        poseidon::flatten_poseidon_hash_value,
+        array::ToField,
+        poseidon::{flatten_poseidon_hash_value, H},
+        types::CBuilder,
         utils::{FromFields, ToFields},
-        C, D,
+        C, D, F,
     };
     use mp2_test::circuit::{run_circuit, UserCircuit};
-    use plonky2::{field::types::Field, plonk::config::Hasher};
-    use rand::{prelude::SliceRandom, thread_rng, Rng};
+    use plonky2::{
+        field::types::Field,
+        iop::{
+            target::Target,
+            witness::{PartialWitness, WitnessWrite},
+        },
+        plonk::config::Hasher,
+    };
+    use rand::{seq::SliceRandom, thread_rng, Rng};
+
+    use crate::{
+        ivc::PublicInputs as OriginalTreePublicInputs,
+        query::{
+            aggregation::{QueryBoundSource, QueryBounds},
+            public_inputs::{
+                    PublicInputs as QueryProofPublicInputs,
+                    QueryPublicInputs,
+                },
+            computational_hash_ids::AggregationOperation,
+            universal_circuit::{
+                universal_circuit_inputs::Placeholders, universal_query_gadget::OutputValues,
+            },
+        },
+        revelation::{
+            revelation_without_results_tree::{
+                RevelationWithoutResultsTreeCircuit, RevelationWithoutResultsTreeWires,
+            },
+            tests::{compute_results_from_query_proof_outputs, TestPlaceholders},
+            PublicInputs, NUM_PREPROCESSING_IO,
+        },
+        test_utils::{random_aggregation_operations, random_original_tree_proof, sample_boundary_rows_for_revelation},
+    };
 
     // L: maximum number of results
     // S: maximum number of items in each result
@@ -381,9 +423,9 @@ mod tests {
     const PP: usize = 20;
 
     // Real number of the placeholders
-    const NUM_PLACEHOLDERS: usize = 5;
+    const NUM_PLACEHOLDERS: usize = 6;
 
-    const QUERY_PI_LEN: usize = crate::query::pi_len::<S>();
+    const QUERY_PI_LEN: usize = QueryProofPublicInputs::<F, S>::total_len();
 
     impl From<&TestPlaceholders<PH, PP>> for RevelationWithoutResultsTreeCircuit<L, S, PH, PP> {
         fn from(test_placeholders: &TestPlaceholders<PH, PP>) -> Self {
@@ -392,15 +434,16 @@ mod tests {
             }
         }
     }
+        
 
     #[derive(Clone, Debug)]
-    struct TestRevelationWithoutResultsTreeCircuit<'a> {
+    struct TestRevelationCircuit<'a> {
         c: RevelationWithoutResultsTreeCircuit<L, S, PH, PP>,
         query_proof: &'a [F],
         original_tree_proof: &'a [F],
     }
 
-    impl UserCircuit<F, D> for TestRevelationWithoutResultsTreeCircuit<'_> {
+    impl UserCircuit<F, D> for TestRevelationCircuit<'_> {
         // Circuit wires + query proof + original tree proof (IVC proof)
         type Wires = (
             RevelationWithoutResultsTreeWires<L, S, PH, PP>,
@@ -433,21 +476,26 @@ mod tests {
         ops: &[F; S],
         test_placeholders: &TestPlaceholders<PH, PP>,
     ) -> Vec<F> {
-        let [mut proof] = random_aggregation_public_inputs(ops);
+        let [mut proof] = QueryProofPublicInputs::sample_from_ops(ops);
 
-        let [count_range, min_query_range, max_query_range, p_hash_range] = [
-            QueryPublicInputs::NumMatching,
-            QueryPublicInputs::MinQuery,
-            QueryPublicInputs::MaxQuery,
-            QueryPublicInputs::PlaceholderHash,
-        ]
-        .map(QueryProofPublicInputs::<F, S>::to_range);
+        let [count_range, min_query_primary, max_query_primary, min_query_secondary, max_query_secondary, p_hash_range, left_row_range, right_row_range] =
+            [
+                QueryPublicInputs::NumMatching,
+                QueryPublicInputs::MinPrimary,
+                QueryPublicInputs::MaxPrimary,
+                QueryPublicInputs::MinSecondary,
+                QueryPublicInputs::MaxSecondary,
+                QueryPublicInputs::PlaceholderHash,
+                QueryPublicInputs::LeftBoundaryRow,
+                QueryPublicInputs::RightBoundaryRow,
+            ]
+            .map(QueryProofPublicInputs::<F, S>::to_range);
 
         // Set the count, minimum, maximum query and the placeholder hash.
         [
             (count_range, vec![entry_count.to_field()]),
-            (min_query_range, test_placeholders.min_query.to_fields()),
-            (max_query_range, test_placeholders.max_query.to_fields()),
+            (min_query_primary, test_placeholders.min_query.to_fields()),
+            (max_query_primary, test_placeholders.max_query.to_fields()),
             (
                 p_hash_range,
                 test_placeholders.query_placeholder_hash.to_fields(),
@@ -456,11 +504,28 @@ mod tests {
         .into_iter()
         .for_each(|(range, fields)| proof[range].copy_from_slice(&fields));
 
+        // Set boundary rows to satisfy constraints for completeness
+        let rng = &mut thread_rng();
+        let min_secondary = U256::from_fields(&proof[min_query_secondary]);
+        let max_secondary = U256::from_fields(&proof[max_query_secondary]);
+        let placeholders =
+            Placeholders::new_empty(test_placeholders.min_query, test_placeholders.max_query);
+        let query_bounds = QueryBounds::new(
+            &placeholders,
+            Some(QueryBoundSource::Constant(min_secondary)),
+            Some(QueryBoundSource::Constant(max_secondary)),
+        )
+        .unwrap();
+        let (left_boundary_row, right_boundary_row) = sample_boundary_rows_for_revelation(&query_bounds, rng);
+
+        proof[left_row_range].copy_from_slice(&left_boundary_row.to_fields());
+        proof[right_row_range].copy_from_slice(&right_boundary_row.to_fields());
+
         proof
     }
 
     /// Utility function for testing the revelation circuit with results tree
-    fn test_revelation_without_results_tree_circuit(ops: &[F; S], entry_count: Option<u32>) {
+    fn test_revelation_batching_circuit(ops: &[F; S], entry_count: Option<u32>) {
         let rng = &mut thread_rng();
 
         // Generate the testing placeholder data.
@@ -476,7 +541,7 @@ mod tests {
         let original_tree_pi = OriginalTreePublicInputs::from_slice(&original_tree_proof);
 
         // Construct the test circuit.
-        let test_circuit = TestRevelationWithoutResultsTreeCircuit {
+        let test_circuit = TestRevelationCircuit {
             c: (&test_placeholders).into(),
             query_proof: &query_proof,
             original_tree_proof: &original_tree_proof,
@@ -486,7 +551,6 @@ mod tests {
         let proof = run_circuit::<F, D, C, _>(test_circuit);
         let pi = PublicInputs::<_, L, S, PH>::from_slice(&proof.public_inputs);
 
-        // Initialize the overflow flag to false.
         let entry_count = query_pi.num_matching_rows();
 
         // Check the public inputs.
@@ -547,57 +611,58 @@ mod tests {
     }
 
     #[test]
-    fn test_revelation_without_results_tree_simple() {
+    fn test_revelation_batching_simple() {
         // Generate the random operations and set the first operation to SUM
         // (not ID which should not be present in the aggregation).
         let mut ops: [_; S] = random_aggregation_operations();
         ops[0] = AggregationOperation::SumOp.to_field();
 
-        test_revelation_without_results_tree_circuit(&ops, None);
+        test_revelation_batching_circuit(&ops, None);
     }
 
     // Test for COUNT operation.
     #[test]
-    fn test_revelation_without_results_tree_for_op_count() {
+    fn test_revelation_batching_for_op_count() {
         // Set the first operation to COUNT.
         let mut ops: [_; S] = random_aggregation_operations();
         ops[0] = AggregationOperation::CountOp.to_field();
 
-        test_revelation_without_results_tree_circuit(&ops, None);
+        test_revelation_batching_circuit(&ops, None);
     }
 
     // Test for AVG operation.
     #[test]
-    fn test_revelation_without_results_tree_for_op_avg() {
+    fn test_revelation_batching_for_op_avg() {
         // Set the first operation to AVG.
         let mut ops: [_; S] = random_aggregation_operations();
         ops[0] = AggregationOperation::AvgOp.to_field();
 
-        test_revelation_without_results_tree_circuit(&ops, None);
+        test_revelation_batching_circuit(&ops, None);
     }
 
     // Test for AVG operation with zero entry count.
     #[test]
-    fn test_revelation_without_results_tree_for_op_avg_with_no_entries() {
+    fn test_revelation_batching_for_op_avg_with_no_entries() {
         // Set the first operation to AVG.
         let mut ops: [_; S] = random_aggregation_operations();
         ops[0] = AggregationOperation::AvgOp.to_field();
 
-        test_revelation_without_results_tree_circuit(&ops, Some(0));
+        test_revelation_batching_circuit(&ops, Some(0));
     }
 
     // Test for no AVG operation with zero entry count.
     #[test]
-    fn test_revelation_without_results_tree_for_no_op_avg_with_no_entries() {
+    fn test_revelation_batching_for_no_op_avg_with_no_entries() {
         // Initialize the all operations to SUM or COUNT (not AVG).
         let mut rng = thread_rng();
-        let ops = std::array::from_fn(|_| {
+        let ops = array::from_fn(|_| {
             [AggregationOperation::SumOp, AggregationOperation::CountOp]
                 .choose(&mut rng)
                 .unwrap()
                 .to_field()
         });
 
-        test_revelation_without_results_tree_circuit(&ops, Some(0));
+        test_revelation_batching_circuit(&ops, Some(0));
     }
 }
+
